feat: Add professional testing and build infrastructure
Introduces a modern development environment to improve the long-term robustness and maintainability of the application.

- Adds Jest as a formal testing framework.
- Implements the first suite of unit tests for the core `calculateStatus` function to verify its behavior under multiple conditions.
- Refactors the monolithic `script.js` into smaller, more manageable modules (`js/task-logic.js` and `js/script.js`) to improve code organization and enable testing.
- Integrates npm for package management, allowing for easy installation of dependencies and the execution of development scripts.
- Updates the README.md with a new "Development" section explaining how to install dependencies and run tests.
