+++ jest.config.cjs
@@ -1,0 +1,5 @@
+module.exports = {
+  transform: {
+    '^.+\\.js$': 'babel-jest',
+  },
+};
